+++ js/components/main.js
@@ -27,7 +27,7 @@ const {NotificationBar} = require('./notificationBar')
 const DownloadsBar = require('../../app/renderer/components/downloadsBar')
 const SiteInfo = require('./siteInfo')
 const BraveryPanel = require('./braveryPanel')
-const ClearBrowsingDataPanel = require('./clearBrowsingDataPanel')
+const ClearBrowsingDataPanel = require('../../app/renderer/components/clearBrowsingDataPanel')
 const ImportBrowserDataPanel = require('../../app/renderer/components/importBrowserDataPanel')
 const WidevinePanel = require('../../app/renderer/components/widevinePanel')
 const AutofillAddressPanel = require('../../app/renderer/components/autofill/autofillAddressPanel')
